gamification worker
