+++ packages/core/src/admin-dashboard.server.ts
@@ -14,8 +14,6 @@ export { listFeatureFlags } from './modules/feature-flag/queries/list-feature-fl
 export { createFeatureFlag } from './modules/feature-flag/use-cases/create-feature-flag';
 export { deleteFeatureFlag } from './modules/feature-flag/use-cases/delete-feature-flag';
 export { editFeatureFlag } from './modules/feature-flag/use-cases/edit-feature-flag';
-export { archiveActivity } from './modules/gamification/use-cases/archive-activity';
-export { editActivity } from './modules/gamification/use-cases/edit-activity';
 export { addIcebreakerPrompt } from './modules/icebreaker/use-cases/add-icebreaker-prompt';
 export { activateMember } from './modules/member/use-cases/activate-member';
 export { updateMemberEmail } from './modules/member/use-cases/update-member-email';
