--- packages/core/src/modules/gamification/use-cases/archive-activity.ts
@@ -1,9 +1,0 @@
-import { db } from '@/infrastructure/database';
-
-export async function archiveActivity(id: string) {
-  await db
-    .updateTable('activities')
-    .set({ deletedAt: new Date() })
-    .where('id', '=', id)
-    .executeTakeFirst();
-}
